+++ src/main/java/com/claySoftware/MathExpAssistant/services/JwtTokenProvider.java
@@ -1,0 +1,49 @@
+package com.claySoftware.MathExpAssistant.services;
+
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.oauth2.core.oidc.user.OidcUser;
+import org.springframework.security.oauth2.core.user.DefaultOAuth2User;
+import org.springframework.stereotype.Service;
+
+import java.util.Date;
+
+@Service
+public class JwtTokenProvider {
+
+    @Value("${jwt.secret}")
+    private String jwtSecret;
+
+    @Value("${jwt.expiration}")
+    private long jwtExpiration;
+
+    public String createToken(Authentication authentication) {
+        String username;
+        String email;
+
+        if (authentication.getPrincipal() instanceof OidcUser) {
+            OidcUser oidcUser = (OidcUser) authentication.getPrincipal();
+            username = oidcUser.getName();
+            email = oidcUser.getEmail();
+        } else if (authentication.getPrincipal() instanceof DefaultOAuth2User) {
+            DefaultOAuth2User oauth2User = (DefaultOAuth2User) authentication.getPrincipal();
+            username = oauth2User.getName();
+            email = oauth2User.getAttribute("email");
+        } else {
+            throw new IllegalArgumentException("Unexpected principal type: " + authentication.getPrincipal().getClass());
+        }
+        return Jwts.builder()
+                .setSubject(username)
+                .claim("email", email)
+                .setIssuedAt(new Date())
+                .setExpiration(new Date((new Date()).getTime() + jwtExpiration))
+                .signWith(SignatureAlgorithm.HS512, jwtSecret)
+                .compact();
+    }
+
+    public String getJwtSecret() {
+        return jwtSecret;
+    }
+}